+++ src/app/basicauction/components/LeftSection.css.ts
@@ -1,4 +1,4 @@
-import { style } from '@vanilla-extract/css';
+import { style, styleVariants } from '@vanilla-extract/css';
 
 import colors from '@/styles/color';
 import typography from '@/styles/typo';
@@ -8,6 +8,8 @@ export const leftSection = style({
   flexDirection: 'column',
   gap: '30px',
 });
+
+// --- Checkbox.css.ts ---
 
 export const label = style({
   ...typography.caption,
@@ -33,4 +35,35 @@ export const checked = style({
 
 export const nonChecked = style({
   color: 'rgb(228, 228, 231)',
+});
+
+// --- SearchBar.css.ts ---
+
+export const searchBar = style({
+  position: 'relative',
+  display: 'flex',
+  gap: '5px',
+  borderBottom: `1px solid ${colors.gray7}`,
+  padding: '0 10px 5px',
+});
+
+const buttonBase = style({
+  display: 'flex',
+  alignItems: 'center',
+});
+
+export const buttonIcon = styleVariants({
+  search: [buttonBase],
+  delete: [
+    buttonBase,
+    {
+      color: colors.gray8,
+    },
+  ],
+});
+
+export const searchInput = style({
+  ...typography.body,
+  border: 'none',
+  outline: 'none',
 });
